0.10.0 hardening: review tech-debt (durability CAS, auth gate, client resilience) (#55)
* chore: bump to 0.10.0-SNAPSHOT for development

Skipping a 0.9.3 release; the durable-completion/gRPC/hermetic-build work
(PR #54) ships in 0.10.0 along with the remaining review tech-debt.


* fix(a2a): 0.10.0 hardening — id equality, shutdown order, normalizer scoping

- #17 JsonRpcId.RawNum strips trailing zeros (value-based numeric id equality).
- #20 interruptAll interrupts fibers before clearing the registry (no
  reconcile-clobber window during shutdown).
- #12 payload normalizer no longer applies A2A shape heuristics inside the
  free-form metadata/data subtrees (won't mutate user objects that happen to
  look like a message/skill/card).
- normalizePartAliases keeps the original file/fileUrl key when no canonical
  content could be produced (was dropping it and leaving an empty part).


* fix(a2a): 0.10.0 hardening — structured client errors + JS resource/overflow

- #10 client surfaces structured JSON-RPC/REST error documents (real code/
  message) on 4xx/5xx instead of opaquing them into invalidAgentResponse.
- SSE client cancels the JS reader on scope close (releases the connection
  instead of lingering to GC).
- request body reader accumulates length in Long so a near-Int.MaxValue cap
  can't wrap and bypass the size check.


* feat(a2a): #3 A2ATaskStore.transformIfNotTerminal CAS for reconcile

Add a compare-and-set primitive to A2ATaskStore: atomically transform a task
only if it exists and is non-terminal, returning the terminal task unchanged
(honoring a concurrent terminal write) or None if absent. Default is best-effort
load/save (source-compatible for existing consumers incl. tjc-agents stores);
the in-memory impl overrides it under its lock; durable backends should override
with real CAS (documented on the trait). reconcileOrphaned now transitions via
it — closing the store-level read-modify-write race (a run completing in the
window is no longer clobbered with failed) and reporting taskNotFound for a
deleted task. +deterministic A2ATaskStoreSpec coverage.


* feat(a2a): #4 gRPC dispatch-layer auth gate

A2AGrpcBinding.dispatch now calls authorizeRequest before routing — the gRPC
transport previously had no auth of its own and relied entirely on each
handler remembering to call it. Made A2ARequestHandler.authorizeRequest public
for transports to use as a single first line of defense (idempotent; handlers
still call it). The full HTTP-pipeline reorder (auth before parse/route to
remove error-code enumeration) is deferred — HTTP already authorizes per-method
so it's defense-in-depth, not a hole.


* Address PR #55 round-3 review: interruptAll deadlock + reconcile + regression tests

Fixes from the claude-review on a2a-0.10.0-hardening:

- #1 (critical) interruptAll deadlock: snapshot the registry (ref.get) and
  interrupt fibers WITHOUT holding the Ref.Synchronized permit, then clear
  (ref.set). The prior modifyZIO impl held the permit across each interrupt,
  but every run's `.ensuring(remove(key))` finalizer re-acquires the same
  permit, so interrupt-vs-finalizer deadlocked on shutdown.
- #2 reconcileOrphaned: re-add the dropped logWarning before the
  transformIfNotTerminal CAS so orphaned-task self-heals stay observable.
- #3 A2ATaskStore: document that durable backends MUST override
  transformIfNotTerminal with a real compare-and-set (in-memory does so under
  its lock; the default is best-effort load-then-save).

Regression tests (all green on JVM + Scala.js):
- A2ARuntimeRegistrySpec: interruptAll terminates within a timeout even when a
  live fiber's finalizer re-enters the registry (would hang on the old impl).
- A2AServerCoreSpec: getTask surfaces taskNotFound when the task is deleted
  during reconcile (CAS returns None) instead of resurrecting a phantom failed
  row.
- A2AClientPayloadNormalizerSpec: the normalizer does not synthesize ids inside
  free-form `metadata` — heuristics stay scoped to structural positions.


* Fix sendMessage follow-up race: join the run fiber before returning

CORE-MULTI-005 (and any sequential follow-up to a still-non-terminal task)
flaked with "Task <id> already has an active run". A stream-ending event
reaches `waitForFinal` from inside the run (bus.finish on closesStream), but
the run's `remove(key)` finalizer runs later, in the forked fiber's `.ensuring`
after `executeRun` returns. So a follow-up send (legal for an input-required /
auth-required task) could `reserve(key)` before the prior run's `remove(key)`,
and prepare rejected it.

Fix: `startExecution` now returns the forked fiber, and the synchronous
`sendMessage` path `.zipLeft(fiber.await)`s after `waitForFinal` — the runtime
entry is guaranteed gone before we return. `fiber.await` (not `join`) is used
so an interrupted run (the cancel path interrupts the fiber) does not propagate
its interruption into the caller; we only need to know it finished.

Verified: agent.jvm.test 6x + agent.js.test 3x green, including the
previously-flaky ACTS multi-turn scenario and the CancelTask live test.


---------

+++ src/js/com/tjclp/scalagent/a2a/A2AServer.scala
@@ -368,9 +368,12 @@ private final class A2AServerLiveImpl(
         if step.selectDynamic("done").asInstanceOf[Boolean] then ZIO.succeed(totalBytes -> chunks)
         else
           val chunk = step.selectDynamic("value")
-          val next  = totalBytes + chunkByteLength(chunk)
-          if next > maxBytes then cancelReader(reader) *> ZIO.fail(A2AHttpBinding.bodySizeExceeded(maxBytes))
-          else readBodyChunks(reader, maxBytes, next, chunks :+ chunk)
+          // Long arithmetic so a near-Int.MaxValue cap can't wrap negative and
+          // silently defeat the size check. `next.toInt` below is safe — we only
+          // recurse when next <= maxBytes (an Int).
+          val next = totalBytes.toLong + chunkByteLength(chunk).toLong
+          if next > maxBytes.toLong then cancelReader(reader) *> ZIO.fail(A2AHttpBinding.bodySizeExceeded(maxBytes))
+          else readBodyChunks(reader, maxBytes, next.toInt, chunks :+ chunk)
       }
 
     private def cancelReader(reader: js.Dynamic): UIO[Unit] =
